Moved widget from app to core
