+++ app/src/main/AndroidManifest.xml
@@ -118,13 +118,13 @@
         </activity>
 
         <service
-            android:name=".service.PlayerWidgetJobService"
+            android:name=".core.service.PlayerWidgetJobService"
             android:permission="android.permission.BIND_JOB_SERVICE"
             android:enabled="true"
             android:exported="false">
         </service>
 
-        <receiver android:name=".receiver.PlayerWidget">
+        <receiver android:name=".core.receiver.PlayerWidget">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE"/>
             </intent-filter>
