+++ core/src/main/res/layout/player_widget.xml
